fix up settings routes
