+++ Composer/packages/client/src/pages/setting/dialog-settings/index.tsx
@@ -10,6 +10,7 @@ import { DefaultButton } from 'office-ui-fabric-react/lib/Button';
 import { ChoiceGroup } from 'office-ui-fabric-react/lib/ChoiceGroup';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
+import { RouteComponentProps } from '@reach/router';
 
 import { StoreContext } from '../../../store';
 import { isAbsHosted } from '../../../utils/envUtil';
@@ -27,7 +28,7 @@ const hostControlLabels = {
   learnMore: formatMessage('Learn more.'),
 };
 
-export const DialogSettings = () => {
+export const DialogSettings: React.FC<RouteComponentProps> = () => {
   const { state, actions } = useContext(StoreContext);
   const { botName, settings: origSettings, botEnvironment, projectId } = state;
   const absHosted = isAbsHosted();
