+++ Composer/packages/client/src/router.tsx
@@ -40,7 +40,7 @@ const Routes = props => {
           />
           <Redirect from="/" to={resolveToBasePath(BASEPATH, 'home')} noThrow />
           <ProjectRouter path="/bot/:projectId">
-            <SettingPage path="settings/:subpage" />
+            <SettingPage path="settings/*" />
             <LUPage path="language-understanding/:dialogId/*" />
             <LGPage path="language-generation/:dialogId/*" />
             <Notifications path="notifications" />
